% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_raster_fx.R
\name{richness_raster}
\alias{richness_raster}
\title{Generate a species richness raster}
\usage{
richness_raster(template.raster, occurrences)
}
\arguments{
\item{template.raster}{The raster that corresponds to the cell numbers in the second column of the occurrences file}

\item{occurrences}{A set of occurrences in "tidy" format: first column is species name, second is raster cells where the species occurs.}
}
\value{
A raster of the same geometry as \code{template.raster}, giving the number of distinct species recorded in each cell. Cells with no occurrences are NA.
}
\description{
richness_raster
}
\examples{
template <- raster::raster(nrows = 10, ncols = 10)

richness <- richness_raster(template.raster = template,
                            occurrences = example_occurrences)

summary(raster::values(richness))
}
