% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_branches_by_traits_fastAnc.R
\name{scale_branches_by_traits_fastAnc}
\alias{scale_branches_by_traits_fastAnc}
\title{Scale the branches of a phylogeny according to phenotypic change.}
\usage{
scale_branches_by_traits_fastAnc(tree, traits, percent = FALSE)
}
\arguments{
\item{tree}{A phylogeny with branch lengths in units of time.}

\item{traits}{a set of trait data where the first column is species name and additional columns are trait data}

\item{percent}{If TRUE, branch lengths returned will reflect percent change, rather than absolute amount of change.}
}
\value{
phylo formate phylogeny
}
\description{
scale_branches_by_traits_fastAnc
}
\examples{
# This function takes a single trait, as a vector named by species
body_mass <- example_traits$body_mass_g
names(body_mass) <- example_traits$species

mass_tree <- scale_branches_by_traits_fastAnc(tree = example_tree,
                                              traits = body_mass)

plot(example_tree$edge.length, mass_tree$edge.length,
     xlab = "time", ylab = "change in body mass")
}
