% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_branches_by_traits_rphylopars.R
\name{scale_branches_by_traits_rphylopars}
\alias{scale_branches_by_traits_rphylopars}
\title{Scale the branches of a phylogeny according to phenotypic change.}
\usage{
scale_branches_by_traits_rphylopars(
  tree,
  traits,
  percent = FALSE,
  pheno_error = NULL
)
}
\arguments{
\item{tree}{A phylogeny with branch lengths in units of time.}

\item{traits}{a set of trait data where the first column is species name and additional columns are trait data}

\item{percent}{If TRUE, branch lengths returned will reflect percent change, rather than absolute amount of change.}

\item{pheno_error}{Logical, or NULL (the default). Controls whether ancestral state reconstruction estimates a within-species (phenotypic) error term. When NULL, the setting is taken from the data: TRUE if at least one species has two or more observations of at least one trait, and FALSE otherwise. Supply TRUE or FALSE to override. See \code{scale_branches_multidimensional}.}
}
\value{
phylo formatted phylogeny
}
\description{
scale_branches_by_traits_rphylopars
}
\note{
This function DOES NOT account for uncertainty in estimated ancestral traits.
}
\examples{
\dontrun{
Write example text
}
}
