% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traittree-package.R
\docType{package}
\name{traittree-package}
\alias{traittree}
\alias{traittree-package}
\title{traittree: Functions for scaling phylogeny by phenotypic traits}
\description{
This package provides a few tools allowing the user to re-scale phylogenies from units of time to units of morphological change.
}
\author{
\strong{Maintainer}: Brian Maitner \email{bmaitner@gmail.com}

}
\keyword{internal}
