% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pd_metric_rasters_fx.R
\name{calculate_pd_metric_rasters}
\alias{calculate_pd_metric_rasters}
\title{Calculate time- and trait-scaled phylogenies and pd/pdi rasters}
\usage{
calculate_pd_metric_rasters(occurrences, phylogeny, traits, template.raster)
}
\arguments{
\item{occurrences}{A set of occurrences in "tidy" format: first column is species name, second is raster cells where the species occurs.}

\item{phylogeny}{A phylogeny with time-scaled branch lengths}

\item{traits}{a set of trait data where the first column is species name and additional columns are trait data}

\item{template.raster}{The raster that corresponds to the cell numbers in the second column of the occurrences file}
}
\value{
List containing 4 rasters and 3 phylogenies (original, trait-scaled and rate-scaled)
}
\description{
calculate_pd_metric_rasters
}
\examples{
\dontrun{
Write example text
}
}
