% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_tree}
\alias{example_tree}
\title{Simulated phylogeny for the package examples}
\format{
An object of class \code{phylo} with 60 tips.
}
\usage{
example_tree
}
\description{
A time-scaled phylogeny of 60 imaginary species, used throughout the examples
and the vignette. Branch lengths are scaled so that the root is at depth 1.
}
\details{
The data are simulated, not observed. They exist so that the examples run
quickly and reproducibly, and should not be used for anything else. The code
that produced them is in \code{data-raw/generate_example_data.R}.
}
\seealso{
\link{example_traits}, \link{example_occurrences}
}
\keyword{datasets}
