% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_raster_fx.R
\name{richness_raster}
\alias{richness_raster}
\title{Generate a species richness raster}
\usage{
richness_raster(template.raster, occurrences)
}
\arguments{
\item{template.raster}{The raster that corresponds to the cell numbers in the second column of the occurrences file}

\item{occurrences}{A set of occurrences in "tidy" format: first column is species name, second is raster cells where the species occurs.}
}
\value{
List containing 4 rasters and 3 phylogenies (original, trait-scaled and rate-scaled)
}
\description{
richness_raster
}
\examples{
\dontrun{
Write example text
}
}
