% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_branches_by_traits_fastAnc.R
\name{scale_branches_by_traits_fastAnc}
\alias{scale_branches_by_traits_fastAnc}
\title{Scale the branches of a phylogeny according to phenotypic change.}
\usage{
scale_branches_by_traits_fastAnc(tree, traits, percent = FALSE)
}
\arguments{
\item{tree}{A phylogeny with branch lengths in units of time.}

\item{traits}{a set of trait data where the first column is species name and additional columns are trait data}

\item{percent}{If TRUE, branch lengths returned will reflect percent change, rather than absolute amount of change.}
}
\value{
phylo formate phylogeny
}
\description{
scale_branches_by_traits_fastAnc
}
\examples{
\dontrun{
Write example text
}
}
