% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/scale_branches_multidimensional_distance_with_variance.R
\name{scale_branches_multidimensional_with_variation}
\alias{scale_branches_multidimensional_with_variation}
\title{Scale the branches of a phylogeny according to phenotypic change, accounting for uncertainty.}
\usage{
scale_branches_multidimensional_with_variation(tree, traits, rate = F)
}
\arguments{
\item{tree}{A phylogeny with branch lengths in units of time.}

\item{traits}{a set of trait data where the first column is species name and additional columns are trait data}

\item{rate}{If TRUE, branch lengths returned will reflect rates of change, rather than absolute amount of change.}
}
\value{
phylo formate phylogeny
}
\description{
scale_branches_multidimensional_with_variation
}
\note{
This function accounts for uncertainty in estimated ancestral traits and can be run multiple times to generate a distribution of phylogenies.
}
\examples{
\dontrun{
Write example text
}
}
